+++ src/BasicInputs/EnumInput.tsx
@@ -4,26 +4,35 @@ import {
   forwardRef,
 } from 'react';
 
+import { humanizeText } from '@/utils';
+
 import { SelectInput } from './SelectInput';
 
-export const ModelInput = forwardRef<
+export const EnumInput = forwardRef<
   ElementRef<typeof SelectInput>,
-  ComponentPropsWithoutRef<typeof SelectInput>
+  Omit<ComponentPropsWithoutRef<typeof SelectInput>, 'value' | 'options'> & {
+    value: string;
+    options: string[];
+  }
 >(({ value, onChange, options, isCreatable, ...selectInputProps }, ref) => {
   const isArray = Array.isArray(value);
+  const stringToOption = (str: string) => ({
+    label: humanizeText(str),
+    value: str,
+  });
   return (
     <SelectInput
       ref={ref}
-      value={value}
-      onChange={onChange}
-      getOptionLabel={(option: typeof value) => option.displayValue}
-      getOptionValue={(option: typeof value) => option.id}
-      options={options}
+      value={
+        value && (isArray ? value.map(stringToOption) : stringToOption(value))
+      }
+      onChange={(newVal, actionMeta) => onChange?.(newVal.value, actionMeta)}
+      options={options?.map(stringToOption)}
       isMulti={isArray}
       clearValue={isArray ? () => [] : undefined}
       isClearable={!isArray}
       closeMenuOnSelect={!isArray}
-      isCreatable={true}
+      isCreatable={isCreatable}
       getNewOptionData={
         isCreatable
           ? (inputValue, optionLabel) => ({